+++ jsk_pcl_ros/include/jsk_pcl_ros/polygon_array_wrapper.h
@@ -1,0 +1,74 @@
+// -*- mode: c++ -*-
+/*********************************************************************
+ * Software License Agreement (BSD License)
+ *
+ *  Copyright (c) 2014, JSK Lab
+ *  All rights reserved.
+ *
+ *  Redistribution and use in source and binary forms, with or without
+ *  modification, are permitted provided that the following conditions
+ *  are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *   * Redistributions in binary form must reproduce the above
+ *     copyright notice, this list of conditions and the following
+ *     disclaimer in the documentation and/o2r other materials provided
+ *     with the distribution.
+ *   * Neither the name of the Willow Garage nor the names of its
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ *  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ *  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ *  FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ *  COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ *  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ *  BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ *  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ *  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ *  LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ *  POSSIBILITY OF SUCH DAMAGE.
+ *********************************************************************/
+
+
+#ifndef JSK_PCL_ROS_POLYGON_ARRAY_WRAPPER_H_
+#define JSK_PCL_ROS_POLYGON_ARRAY_WRAPPER_H_
+
+#include <pcl_ros/pcl_nodelet.h>
+
+#include <message_filters/subscriber.h>
+#include <message_filters/time_synchronizer.h>
+#include <message_filters/synchronizer.h>
+
+#include "jsk_pcl_ros/pcl_conversion_util.h"
+
+#include <geometry_msgs/PolygonStamped.h>
+#include <jsk_pcl_ros/PolygonArray.h>
+#include <jsk_pcl_ros/ModelCoefficientsArray.h>
+
+namespace jsk_pcl_ros
+{
+  class PolygonArrayWrapper: public pcl_ros::PCLNodelet
+  {
+  public:
+    typedef message_filters::sync_policies::ExactTime<
+      geometry_msgs::PolygonStamped, PCLModelCoefficientMsg> SyncPolicy;
+
+  protected:
+    virtual void onInit();
+    virtual void wrap(const geometry_msgs::PolygonStamped::ConstPtr& polygon,
+                      const PCLModelCoefficientMsg::ConstPtr& coefficients);
+    boost::shared_ptr<message_filters::Synchronizer<SyncPolicy> >sync_;
+    message_filters::Subscriber<geometry_msgs::PolygonStamped> sub_polygon_;
+    message_filters::Subscriber<PCLModelCoefficientMsg> sub_coefficients_;
+    ros::Publisher pub_polygon_array_;
+    ros::Publisher pub_coefficients_array_;
+  private:
+    
+  };
+}
+
+#endif